+++ hubdle-project-gradle-plugin/main/kotlin/com/javiersc/hubdle/project/extensions/config/documentation/site/HubdleConfigDocumentationSiteExtension.kt
@@ -1,6 +1,9 @@
 package com.javiersc.hubdle.project.extensions.config.documentation.site
 
 import com.javiersc.gradle.project.extensions.isRootProject
+import com.javiersc.gradle.properties.extensions.getStringPropertyOrNull
+import com.javiersc.hubdle.project.HubdleProperty.Analysis.Qodana
+import com.javiersc.hubdle.project.HubdleProperty.Analysis.Sonar
 import com.javiersc.hubdle.project.extensions.HubdleDslMarker
 import com.javiersc.hubdle.project.extensions._internal.ApplicablePlugin.Scope
 import com.javiersc.hubdle.project.extensions._internal.Configurable.Priority
@@ -69,7 +72,9 @@ constructor(
                 PrebuildSiteTask.register(project) {
                     projectsInfo.set(project.projectsInfo)
                     qodana.set(analysis.qodana)
+                    qodanaProjectKey.set(provider { getStringPropertyOrNull(Qodana.projectKey) })
                     sonar.set(analysis.sonar)
+                    sonarProjectId.set(provider { getStringPropertyOrNull(Sonar.projectKey) })
                 }
 
             BuildSiteTask.register(project, preBuildDocsTask)