+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 
 ### Updated
 
+- `gradle -> 8.0.2`
 - `com.javiersc.semver:semver-gradle-plugin -> 0.5.0-alpha.1`
 - `com.javiersc.hubdle:com.javiersc.hubdle.gradle.plugin -> 0.5.0-alpha.2`